requirements.txt
ultralytics
supervision
streamlit
opencv-python-headless
numpy
Pillow
torch
torchvision
streamlit-webrtc
